+++ CHANGES.md
@@ -5,8 +5,10 @@ LUCA Changes
 -----
 _Released October ??, 2013_
 
-* Better display of datestamp
+* Better display of date
 * Import `argparse` module for command-line arguments
+* Remove spaces in folder names
+* Added command-line option (`-d`, `--date`) to append folders with Creation date, formatting as `YYYY-DD-MM`
 
 1.0.2
 -----